Separate config from sender service
