+++ cmd/gses2-app/main.go
@@ -8,13 +8,19 @@ import (
 
 	"gses2-app/internal/controllers"
 	"gses2-app/internal/services"
+	"gses2-app/pkg/config"
 )
 
 func main() {
+	err := config.Load("./config.yaml")
+	if err != nil {
+		log.Fatal(err)
+	}
+
 	httpClient := &http.Client{Timeout: time.Second * 10}
 	exchangeRateService := services.NewExchangeRateService(httpClient)
 	emailSubscriptionService := services.NewEmailSubscriptionService("./storage.csv")
-	emailSenderService := services.NewEmailSenderService("./config.yaml")
+	emailSenderService := services.NewEmailSenderService()
 
 	controller := controllers.NewAppController(
 		exchangeRateService,
@@ -27,8 +33,5 @@ func main() {
 	http.HandleFunc("/api/sendEmails", controller.SendEmails)
 
 	fmt.Println("Starting server on port 8080")
-	err := http.ListenAndServe(":8080", nil)
-	if err != nil {
-		log.Fatal("ListenAndServe: ", err)
-	}
+	log.Fatal(http.ListenAndServe(":8080", nil))
 }
