+++ pkg/config/config.go
@@ -1,0 +1,23 @@
+package config
+
+import (
+	"os"
+
+	yaml "gopkg.in/yaml.v2"
+)
+
+var configuration Config
+
+func Load(filename string) error {
+	data, err := os.ReadFile(filename)
+	if err != nil {
+		return err
+	}
+
+	err = yaml.Unmarshal(data, &configuration)
+	return err
+}
+
+func Current() Config {
+	return configuration
+}
